requirements.txt
Flask==2.0.1
Flask-Login==0.5.0
pymongo==3.12.0
python-dotenv==0.19.1
